requirements.txt
# ALL Detection System - Python Requirements

## Core Dependencies
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
scipy>=1.7.0

## Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
tensorflow>=2.10.0  # Desktop training / TFLite export

## GUI
customtkinter>=5.2.0
Pillow>=9.0.0

## AI / LLM
ollama>=0.1.0

## Data & Visualization
matplotlib>=3.4.0
pandas>=1.3.0
seaborn>=0.11.0
albumentations>=1.3.0

## XAI
tf-keras-vis>=0.8.0

## Utilities
tqdm>=4.62.0
pyyaml>=6.0
rich>=13.0.0
psutil>=5.9.0
